moved table modules to a single modules.py file and organized imports
