+++ db/discord_db/db.py
@@ -1,14 +1,10 @@
-from base import Base
-from user import User
-from channel import Channel
-from message import Message
-
 import os
 
 from dotenv import load_dotenv
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
+from modules import Base
 
 load_dotenv()
